ios: a 64-bit finger id in an int is why the map would not drag
SDL_FingerID is Sint64 and SDL's iOS backend derives it from a UITouch
pointer (SDL_uikitview.m: (SDL_FingerID)((size_t)touch)), which on arm64
is far above INT32_MAX. translateSdlEvents stored it in an int, so the
comparison deciding IsPrimary weighed a truncated value against the real
id and lost, for a single uncomplicated finger, every time.

forms/control.cpp:212 reads "if (e->finger().IsPrimary || 1)", so buttons
worked anyway and hid the fault; tileview.cpp:103 checks IsPrimary
honestly, so the city and battle maps never panned. That asymmetry is
exactly what a player reports as "you can't touch to drag".

The id was also re-derived every poll from SDL_GetTouchFinger(dev, 0), a
live array index rather than a stable identity, so primacy could pass to
another finger mid-gesture. The first finger down now claims it and holds
it until it lifts, and FingerEvent::Id is widened so the same truncation
cannot return through another door.

Desktop is untouched: these cases run only for SDL_FINGER* events, behind
optionEnableTouchEvents, which defaults false off iOS and Android.

Verified by source, not by observation: driving touches on the simulator
needs input injection that was unavailable, so the drag itself is
unconfirmed on a device.

+++ framework/event.h
@@ -4,6 +4,8 @@
 #include "library/sp.h"
 #include "library/strings.h"
 
+#include <cstdint>
+
 namespace OpenApoc
 {
 
@@ -77,8 +79,9 @@ typedef struct FrameworkFingerEvent
 	int Y;
 	int DeltaX;
 	int DeltaY;
-	// Touch ID (system-specified)
-	int Id;
+	// Touch ID (system-specified). 64-bit: iOS derives it from a UITouch pointer,
+	// which does not fit in an int.
+	int64_t Id;
 	// Should this be considered a "primary" touch? (first finger?)
 	bool IsPrimary;
 } FrameworkFingerEvent;
